+++ nexus/social_media/services.py
@@ -1,9 +1,11 @@
+# -*- coding: utf-8 -*-
 # Standard Library
 from datetime import datetime
 
 # Third Party Stuff
 import facebook
 import requests
+import tweepy
 from django.conf import settings
 from django.utils import timezone
 from rest_framework import status
@@ -23,27 +25,27 @@ def upload_image_to_linkedin(author, headers, post, linkedin_auth):
 
     :param post: Post model instance.
 
-    :param linkedin: LINKEDIN_AUTH object taken from settings.
+    :param linkedin_auth: LINKEDIN_AUTH object taken from settings.
 
-    :returns: A valid HTTP response.
+    :returns: value of asset
 
     """
     post_data_assets = {
-        "registerUploadRequest": {
-            "recipes": [
-                "urn:li:digitalmediaRecipe:feedshare-image"
+        'registerUploadRequest': {
+            'recipes': [
+                'urn:li:digitalmediaRecipe:feedshare-image'
             ],
-            "owner": author,
-            "serviceRelationships": [
+            'owner': author,
+            'serviceRelationships': [
                 {
-                    "relationshipType": "OWNER",
-                    "identifier": "urn:li:userGeneratedContent"
+                    'relationshipType': 'OWNER',
+                    'identifier': 'urn:li:userGeneratedContent'
                 }
             ]
         }
     }
 
-    api_url_assets = f"{settings.LINKEDIN_API_URL_BASE}assets?action=registerUpload"
+    api_url_assets = f'{settings.LINKEDIN_API_URL_BASE}assets?action=registerUpload'
 
     response_assets = requests.post(api_url_assets, json=post_data_assets,
                                     headers=headers)
@@ -83,23 +85,23 @@ def publish_on_linkedin(post_id):
                'Content-Type': 'application/json',
                'Authorization': f"Bearer {linkedin_auth['access_token']}"}
 
-    api_url_ugc = f"{settings.LINKEDIN_API_URL_BASE}ugcPosts"
+    api_url_ugc = f'{settings.LINKEDIN_API_URL_BASE}ugcPosts'
 
     post = Post.objects.get(pk=post_id)
 
     post_data = {
-        "author": author,
-        "lifecycleState": "PUBLISHED",
-        "specificContent": {
-            "com.linkedin.ugc.ShareContent": {
-                "shareCommentary": {
-                    "text": post.text
+        'author': author,
+        'lifecycleState': 'PUBLISHED',
+        'specificContent': {
+            'com.linkedin.ugc.ShareContent': {
+                'shareCommentary': {
+                    'text': post.text
                 },
-                "shareMediaCategory": "NONE"
+                'shareMediaCategory': 'NONE'
             },
         },
-        "visibility": {
-            "com.linkedin.ugc.MemberNetworkVisibility": "PUBLIC"
+        'visibility': {
+            'com.linkedin.ugc.MemberNetworkVisibility': 'PUBLIC'
         },
     }
 
@@ -110,13 +112,13 @@ def publish_on_linkedin(post_id):
         share_content = specific_content.get('com.linkedin.ugc.ShareContent')
 
         image_media = [{
-            "status": "READY",
-            "description": {
-                "text": ""
+            'status': 'READY',
+            'description': {
+                'text': ''
             },
-            "media": asset,
-            "title": {
-                "text": ""
+            'media': asset,
+            'title': {
+                'text': ''
             }
         }]
         share_content.update(shareMediaCategory='IMAGE',
@@ -161,13 +163,54 @@ def appropriate_response_action(response):
             raise exc.RequestValidationError(error)
 
 
+def get_twitter_api_object(TWITTER_OAUTH):
+    """Function to generate a twitter API object.
+
+    :param TWITTER_OAUTH: A dictionary having essential twitter oauth tokens viz.
+    consumer_key, consumer_secret, access_key, access_secret.
+
+    :returns: twitter API object.
+
+    :raises WrongArguments: Exception, when invalid twitter oauth token(s) are provided.
+
+    """
+    try:
+        auth = tweepy.OAuthHandler(TWITTER_OAUTH['consumer_key'], TWITTER_OAUTH['consumer_secret'])
+        auth.set_access_token(TWITTER_OAUTH['access_key'], TWITTER_OAUTH['access_secret'])
+        twitter_api = tweepy.API(auth)
+        return twitter_api
+    except tweepy.error.TweepError as excinfo:
+        raise exc.WrongArguments(str(excinfo))
+
+
+def publish_on_twitter(post_id):
+    """Function to post on twitter.
+
+    :param post_id: UUID of the post instance to be posted.
+
+    :raises BadRequest: Exception, when unable to post to twitter using tweepy.
+
+    """
+    post = Post.objects.get(pk=post_id)
+    twitter_api = get_twitter_api_object(settings.TWITTER_OAUTH)
+
+    try:
+        if post.image:
+            filename = post.image.file.name
+            twitter_api.update_with_media(filename=filename, status=post.text, file=post.image)
+        elif post.text:
+            twitter_api.update_status(status=post.text)
+    except tweepy.error.TweepError as excinfo:
+        raise exc.BadRequest(str(excinfo))
+
+
 def get_fb_page_graph():
     graph = facebook.GraphAPI(settings.FB_USER_ACCESS_TOKEN)
     pages = graph.get_object('me/accounts')['data']
     page_access_token = None
     page_list = list(filter(lambda page: page['id'] == settings.FB_PAGE_ID, pages))
     if not page_list:
-        raise exc.WrongArguments("Facebook Page access token could not be found")
+        raise exc.WrongArguments('Facebook Page access token could not be found')
     page_access_token = page_list[0]['access_token']
     page_graph = facebook.GraphAPI(page_access_token)
     return page_graph
@@ -184,32 +227,26 @@ def publish_on_facebook(post_id):
             page_graph.put_photo(image=post.image.file.open('rb'))
     elif post.text:
         page_graph.put_object(
-            parent_object=settings.FB_PAGE_ID, connection_name='feed',
-            message=post.text
+            parent_object=settings.FB_PAGE_ID, connection_name='feed', message=post.text
         )
 
 
 def publish_on_social_media():
-    if settings.LIMIT_POSTS is True and int(settings.MAX_POSTS_AT_ONCE) > 0:
-        posts = Post.objects.filter(
-            is_approved=True, is_posted=False,
-            scheduled_time__lte=datetime.now()
-        )[:int(settings.MAX_POSTS_AT_ONCE)]
-    else:
-        posts = Post.objects.filter(
-            is_approved=True, is_posted=False,
-            scheduled_time__lte=datetime.now()
-        )
+    posts = Post.objects.filter(is_approved=True, is_posted=False, scheduled_time__lte=datetime.now())
 
-    # Before bulk update, saving the IDs of posts along with there publishing platforms.
-    post_platform = {}
-    for post in posts:
-        post_platform.update({post.id: post.posted_at})
+    if settings.LIMIT_POSTS is True and int(settings.MAX_POSTS_AT_ONCE) > 0:
+        posts = posts[:int(settings.MAX_POSTS_AT_ONCE)]
+
+    # Before performing the bulk update, here we are saving the IDs of posts along with there publishing platforms.
+    # Because this queryset "posts" will get empty, after running the update query.
+    post_id_and_platform = list(posts.values('id', 'posted_at'))
 
     Post.objects.filter(id__in=posts).update(is_posted=True, posted_time=timezone.now())
 
-    for post_id in post_platform:
-        if post_platform[post_id] == 'fb':
-            tasks.publish_on_facebook_task.delay(post_id)
-        elif post_platform[post_id] == 'linkedin':
-            tasks.publish_on_linkedin_task.delay(post_id)
+    for post in post_id_and_platform:
+        if post['posted_at'] == 'fb':
+            tasks.publish_on_facebook_task.delay(post['id'])
+        elif post['posted_at'] == 'twitter':
+            tasks.publish_on_twitter_task.delay(post['id'])
+        elif post['posted_at'] == 'linkedin':
+            tasks.publish_on_linkedin_task.delay(post['id'])